requirements/dev.txt
-r base.txt
django-debug-toolbar
Fabric

djangorestframework==3.4.6
djangorestframework-expiring-authtoken==0.1.4
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.1.2
django-appconf==1.0.2
django-rest-framework-social-oauth2==1.0.8
enum34==1.1.6
Faker==0.7.17
ipython==5.4.1
ipython-genutils==0.2.0
libsass==0.13.2
MarkupSafe==1.0
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyMySQL==0.7.11
rcssmin==1.0.6
rjsmin==1.0.12
scandir==1.5
stripe==1.73.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
MySQL-python==1.2.5
twilio==6.0.0
celery[redis]==3.1.18
mock==1.3.0
drf-writable-nested==0.5.1

